Feat/validacao dados metodos

+++ app/Http/Controllers/MarcaController.php
@@ -24,9 +24,10 @@ class MarcaController extends Controller
      */
     public function index(Marca $marca)
     {
+        // Listar todos os regitros;
         $marcas = $this->marca->all(); // usando a instância do objeto
         // $marcas =  Marca::all(); // não fazemos a instância do objeto
-        return $marcas;
+        return response()->json($marcas,200);
     }
 
     /**
@@ -41,10 +42,25 @@ class MarcaController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
+    { 
+        // Criar um novo registro;
         // $marca = Marca::create($request->all()); // não fazemos a instância do objeto
+        // nome
+        // imagem
+
+        $regras = [
+            'nome' => 'required|unique:marcas',
+            'imagem' => 'required'
+        ];
+
+        $feedback = [
+            'required' => 'O campo :attribbute é obrigatório',
+            'nome.unique' => 'O nome da marca já existe'
+        ];
+
+        $request->validate($regras, $feedback);
         $marca = $this->marca->create($request->all()); // usando a instância do objeto
-        return $marca; 
+        return response()->json($marca, 201); 
     }
 
     /**
@@ -52,8 +68,12 @@ class MarcaController extends Controller
      */
     public function show($id)
     {
+        // Mostrar os dados de um registro esoecífico
         $marca = $this->marca->find($id); // usando a instância do objeto
-        return $marca; // sugestão de tipo
+        if($marca == null){
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404); // json
+        }
+        return response()->json($marca,200); // sugestão de tipo
     }
 
     /**
@@ -69,6 +89,8 @@ class MarcaController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // Atualizar dados de um registro existente;
+
         /*print_r($request->all()); // os dados atualizados
         echo '<hr>';
         print_r($marca->getAttributes()); // os dados antigos
@@ -76,8 +98,13 @@ class MarcaController extends Controller
 
         //$marca->update($request->all());
         $marca = $this->marca->find($id); // usando a instância do objeto
+
+        if($marca == null){
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'],404);
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca,200); // sugestão de tipo
+
 
         //Para que serve o find?
         //O método find() é usado para buscar um registro específico no banco de dados com base em seu ID. Ele retorna uma instância do modelo correspondente ao registro encontrado ou null se nenhum registro for encontrado com o ID fornecido. No exemplo acima, $this->marca->find($id) busca a marca com o ID especificado e retorna a instância do modelo Marca correspondente a esse registro.
@@ -94,10 +121,13 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id); // usando a instância do objeto
+        if($marca == null){
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'],404);
+        } 
         $marca->delete();
         //print_r($marca->getAttributes()); // os dados antigos
         // getAttributes() é um método do Eloquent que retorna um array com os atributos atuais/alterados do modelo, ou seja, os dados do registro no banco de dados.
-        return ['msg' => 'Marca deletada com sucesso!'];
+        return response()->json(['msg' => 'Marca deletada com sucesso!'],200);
     }
 }
 
@@ -116,9 +146,32 @@ O que é Type Hinting?
 Type Hinting é um recurso do PHP que permite especificar o 
 tipo de dado esperado para um parâmetro ou o tipo 
 de dado retornado por uma função ou método. 
-Isso ajuda a garantir que os dados sejam do tipo correto, 
+Isso ajuda a garantir que os dados sejam do tipo cor?reto, 
 evitando erros e melhorando a legibilidade do código.
 No exemplo acima, a sugestão de tipo "Marca" indica que
 o parâmetro $marca deve ser uma instância da classe Marca, e o
 retorno da função show() também deve ser do tipo Marca.
+
+
+
+O que é o Accept?
+É um cabeçalho(header) da requisição HTTP que informa ao servidor qual 
+formato de resposta o cliente deseja receber. 
+Ex:
+GET /produtos
+Accept: application/json
+{
+  "id": 1,
+  "nome": "Coca-Cola"
+}
+
+
+
+GET /produtos
+Accept: application/xml
+<produto>
+  <id>1</id>
+  <nome>Coca-Cola</nome>
+</produto>
+
 */
